requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
numpy
plotly
streamlit
requests
networkx
graphviz
lxml
groq
google-genai
alpaca-trade-api
langchain
langchain-core
python-dotenv
httpx
apscheduler
